+++ forge-1.20.1/src/main/java/net/tysontheember/remapids/forge/mixin/MappedRegistryMixin.java
@@ -8,6 +8,7 @@ import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.tysontheember.remapids.api.RemapType;
 import net.tysontheember.remapids.core.RemapState;
+import net.tysontheember.remapids.forge.ForgePlatformHelper;
 import org.slf4j.Logger;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -41,6 +42,10 @@ public abstract class MappedRegistryMixin<T> {
 
     @Inject(method = "freeze", at = @At("HEAD"))
     private void remapids$beforeFreeze(CallbackInfoReturnable<Registry<T>> cir) {
+        if (RemapState.hasPending()) {
+            RemapState.finalizeIfPending(ForgePlatformHelper.getAllRegistryIds());
+        }
+
         RemapType type = registryKeyToRemapType(this.key());
         if (type == null) return;